+++ tests/js/index.js
@@ -18,7 +18,10 @@
 
 'use strict';
 
-var Realm = require('realm');
+const Realm = require('realm');
+
+// FIXME: sync testing needs to be updated for ROS 2.0
+global.enableSyncTests = Realm.Sync && false;
 
 var TESTS = {
     ListTests: require('./list-tests'),
@@ -37,14 +40,15 @@ if (!(typeof process === 'object' && process.platform === 'win32')) {
 }
 
 // If sync is enabled, run the sync tests
-if (Realm.Sync) {
+if (global.enableSyncTests) {
     TESTS.UserTests = require('./user-tests');
     TESTS.SessionTests = require('./session-tests');
 
     // FIXME: Permission tests currently fail in chrome debugging mode.
     if (typeof navigator === 'undefined' ||
-        !/Chrome/.test(navigator.userAgent)) { // eslint-disable-line no-undef
-      TESTS.PermissionTests = require('./permission-tests');
+      !/Chrome/.test(navigator.userAgent)) { // eslint-disable-line no-undef
+      
+     TESTS.PermissionTests = require('./permission-tests');
     }
 }
 
@@ -53,7 +57,7 @@ function node_require(module) { return require(module); }
 // If on node, run the async tests
 const isNodeProcess = typeof process === 'object' && process + '' === '[object process]';
 if (isNodeProcess) {
-    //TESTS.AsyncTests = node_require('./async-tests');
+    TESTS.AsyncTests = node_require('./async-tests');
 }
 
 var SPECIAL_METHODS = {
@@ -82,7 +86,7 @@ exports.registerTests = function(tests) {
 };
 
 exports.prepare = function(done) {
-    if (!Realm.Sync || !isNodeProcess || global.testAdminUserInfo) {
+    if (!global.enableSyncTests || !isNodeProcess || global.testAdminUserInfo) {
         done();
         return;
     }