+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-2.0.0 Release notes
+2.0.0-rc10 Release notes (2017-9-19)
 =============================================================
 ### Breaking changes
-* Updating core, sync, object store.
+* Updating core (3.2.1), sync (2.0.0-rc23 - packaged under 2.0.0-rc22), object store.
 
 ### Enhancements
 * Throw exception with recovery path when upgrading from 1.x to 2.x.
+* Add support for sorting Lists and Results on values from linked objects.
 
 ### Bug fixes
 * None